Pizzas Display at Basket
Now when we finish making up a pizza, it shows up in the basket.

Size, price and toppings are all properly displayed.

I am using a bool property to indicate whether a pizza has been finished or has been discarded. Discarded pizzas remain in the order at this poing but doesn't appear in the Basket. However, when implementing the OrderDone functionality, we'll only leave the completed ones remain.