ChooseSide/AddSide
Now we have two controller actions. One ChooseSide returning the view for choosing side and with the model that enables us to pass the content of our order to _BasketPartial; and an AddSide action that adds the side to the order and passes the orderID back to the ChooseSide to return back the same view with the newly populated entity.