ChooseDrink/AddDrink Actions, ChooseDrink View, _Basket Changes
To action methods have been implemented in the BlazinController to enable display and submit of drinks to the Order. Very similar to how I have handled the same functions for Side objects.

I have made changes to the _BasketPartial view to display the drinks exactly the same way as it displays sides, but under its own seperate header, of course.

One problem I now have both with drinks and sides is the I can only add each sides/drink to the order once. It's because the drinks and sides exist as database entities. I am not sure if I want to change this approach just yet. For now I will try to find a way to work around this problem.

After that I will seek to implement payment functionality or at least quasi-functionality.