ReviewOrder/_OrderPartial
Now we have a ReviewOrder view that lets the user review the order in detail before proceeding. I might add the delete functionality here to keep things tidier in the Order view.

We don't have a finalize button yet, which will lead to the Payment view that is to be created. Neither do we have the remove item function.