FinalizeOrder Done
I have fixed some issues with the relationships between the database entities. Since we have classes such as DrinksPerOrder and SidesPerOrder that goes inbetween the Drinks and the Order, and the Sides and the Order, we do not need a direct relationship betwen Drinks/Sides and Order. I have deleted the relationships between Drinks/Sides and order to prevent the creation of unnecessary tables.

Some minor style changes have also been made. Such as the display of Sides and Drinks in the basket and in the Edit/Finalize section. Next is to add a page with form for payment details.