ThankYou Action / Authentication
Authentication is now achieved in the ThankYou action of the BlazinController, where it is checked whether the user is a signed in member, and if so, it is checked whether there is a corresponding Member entity in the database. And an appropriate action is choosed according to the result of these conditions.